Hamfisted attempt at romsimm support.
Start maciici with -cart <romfile> and it will replace the
builtin ROM.  However, the emulated ROM overlay does not appear
to be correct.
The maciici jumps to the address stored at address 0x00000004
when booting.  The maciici ROM contains address 4080002A at that
location, so disabling the ROM overlay after that initial access
works.  However, the maciisi ROM, which boots the IIci (hardware)
just fine, contains address 0000002A, so disabling the ROM overlay
after the initial access doesn't work.  Using a IIsi ROM on the
emulated IIci fails to boot after jumping to 0x2A, which is then
all 0's because the overlay is disabled.
To workaround this, when using a rom simm image, ensure the
address at offset 4 is in the 40800000 range.